+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 4.3.2
 + [x] Slider: fix vertical slider marker position when slider is not visible, thx to [thinkcpu](https://github.com/thinkcpu), PR #1417
-+ [x] Slider: improve events 
++ [x] Slider: improve events
++ [x] HotKey: extended anchor work with `data-hotkey` and attr `href`, issue #1420  
 
 ### 4.3.1
 + [x] Interop: fix interop with Metro4 and jQuery, issue #1408